# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/wallets": {
      "get": {
        "operationId": "WalletsController_list",
        "summary": "List all wallets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns user wallets",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440010",
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "asset": "USD",
                      "balance": "10000.50",
                      "balanceAvailable": "10000.50",
                      "balanceLocked": "0"
                    },
                    {
                      "id": "bb0e8400-e29b-41d4-a716-446655440011",
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "asset": "BTC",
                      "balance": "0.15",
                      "balanceAvailable": "0.14",
                      "balanceLocked": "0.01"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/wallets/{asset}": {
      "get": {
        "operationId": "WalletsController_get",
        "summary": "Get wallet by asset",
        "parameters": [
          {
            "name": "asset",
            "required": true,
            "in": "path",
            "description": "Asset symbol (USD, BTC, ETH)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns wallet",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "aa0e8400-e29b-41d4-a716-446655440010",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "asset": "USD",
                    "balance": "10000.50",
                    "balanceAvailable": "10000.50",
                    "balanceLocked": "0"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/wallets/deposit": {
      "post": {
        "operationId": "WalletsController_deposit",
        "summary": "Deposit (training mode)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training credit: deposit row + updated balance + ledger entry",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "deposit": {
                      "id": "dd0e8400-e29b-41d4-a716-446655440030",
                      "depositType": "fiat",
                      "fiatCurrency": "USD",
                      "amount": "100.00",
                      "status": "completed",
                      "completedAt": "2026-04-09T12:00:00.000Z"
                    },
                    "balance": {
                      "id": "aa0e8400-e29b-41d4-a716-446655440010",
                      "asset": "USD",
                      "balance": "10100.50",
                      "balanceAvailable": "10100.50",
                      "balanceLocked": "0"
                    },
                    "transaction": {
                      "id": "bt-1",
                      "type": "deposit",
                      "amount": "100",
                      "balanceBefore": "10000.5",
                      "balanceAfter": "10100.5",
                      "refType": "deposit_fiat",
                      "refId": "dd0e8400-e29b-41d4-a716-446655440030",
                      "createdAt": "2026-04-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/wallets/withdraw": {
      "post": {
        "operationId": "WalletsController_withdraw",
        "summary": "Withdraw",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated balance and withdraw ledger entry",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "balance": {
                      "id": "aa0e8400-e29b-41d4-a716-446655440010",
                      "asset": "USD",
                      "balance": "9900.00",
                      "balanceAvailable": "9900.00",
                      "balanceLocked": "0"
                    },
                    "transaction": {
                      "id": "bt-wd-2",
                      "type": "withdraw",
                      "amount": "-100",
                      "balanceBefore": "10000",
                      "balanceAfter": "9900",
                      "refType": null,
                      "refId": null,
                      "createdAt": "2026-04-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/users/{userId}/wallets/{asset}": {
      "get": {
        "operationId": "AdminWalletsController_getWallet",
        "summary": "Get user wallet by asset (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asset",
            "required": true,
            "in": "path",
            "description": "Asset symbol (USD, BTC, ETH)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns wallet",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "aa0e8400-e29b-41d4-a716-446655440010",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "asset": "USD",
                    "balance": "10000.50",
                    "balanceAvailable": "10000.50",
                    "balanceLocked": "0"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Wallet not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Test Exchange API",
    "description": "Crypto exchange training platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DepositDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "enum": [
              "USD",
              "EUR",
              "BTC",
              "ETH"
            ],
            "example": "USD"
          },
          "amount": {
            "type": "number",
            "example": 100,
            "minimum": 0,
            "maximum": 1000000000,
            "description": "Positive amount; max 1,000,000,000 (aligned with dashboard client validation)."
          }
        },
        "required": [
          "asset",
          "amount"
        ]
      },
...
    }
  }
}